backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
passlib==1.7.4
bcrypt==4.2.1
pyjwt==2.10.1
pymysql==1.1.1
sqlalchemy==2.0.37
httpx==0.28.1
cryptography==44.0.0
faker==33.3.1
python-multipart== 0.0.20